+++ servers/src/lib.rs
@@ -21,8 +21,10 @@
 #![deny(unused_mut)]
 #![warn(missing_docs)]
 
+extern crate bufstream;
 extern crate hyper;
 extern crate itertools;
+extern crate jsonrpc_core;
 extern crate rand;
 extern crate serde;
 #[macro_use]
@@ -31,8 +33,6 @@ extern crate serde_json;
 #[macro_use]
 extern crate slog;
 extern crate time;
-extern crate jsonrpc_core;
-extern crate bufstream;
 
 extern crate grin_api as api;
 extern crate grin_chain as chain;
